feature
bigDecimal conversions

+++ src/main/java/MSGUserService/models/dtos/UserDto.java
@@ -5,19 +5,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserDto {
 
-    private long id;
+    private BigDecimal id;
 
     private String username;
 
     private String passwordDigest;
 
-    private long userCode;
+    private BigDecimal userCode;
 
     private String email;
